Updated package and application
In preparing to make this available via NPM, I have update the
package.json and also updated the code to deal with multiple
requests and better handle certain exceptions.

This is still an early version, so there may still be other
issues. Oh, and we really need to add some tests.